Arrumando pastas
Correção de Diretórios dos últimos arquivos.

+++ src/ProjetoGeral01/Calculadora.java
@@ -1,5 +1,5 @@
 package ProjetoGeral01;
 
 public class Calculadora {
-    private int num1, num2, resultado;
+    private Double num1, num2, resultado;
 }
